Accordions fully (kind of) functional
The accordions actually work now (Ryan, why didn't you include a script
tag for bootstrap.js!?!?!?!?!?!?) except for making them focusin/out
collapsing. However, that requires a bunch of java tomfoolery that I
don't know how to do, essentially it just needs a focus in function that
has the same effect as clicking on one of the <a>s that collapse or
expand the accordions, and an opposite focusout. That's really all
that's left in the accordions aside from the adding function. Oh, and
why is there a button to decide block? I assume most teachers classes
don't change block each assignment sheet...